Try to get a more meaningful diff.
